# app/Models/Resource.php
<?php

namespace App\Models;

use Illuminate\Database\Eloquent\Factories\HasFactory;
use Illuminate\Database\Eloquent\Model;
use Illuminate\Database\Eloquent\Relations\BelongsTo;
use Illuminate\Database\Eloquent\Relations\HasMany;

class Resource extends Model
{
    /** @use HasFactory<\Database\Factories\ResourceFactory> */
    use HasFactory;

    protected $fillable = [
        'name', 'description', 'capacity'
    ];

    /**
     * Получить все бронирования для ресурса.
     */
    public function bookings(): HasMany
    {
        return $this->hasMany(Booking::class);
    }

    /**
     * Получить пользователя, создавшего ресурс.
     */
    public function user(): BelongsTo
    {
        return $this->belongsTo(User::class);
    }

}